Bug 1055794 - Allows explicit setting of an interpreter for running tasks
On non-posix platforms (such as Windows) hashbangs are not necessarily obeyed. As such,
in order to launch runner from a native shell (such as cmd.exe) it is necessary
to set the interpreter explicitly in runner's task Popen command. These changes
allow each task to be configured with an explicit interpreter, or, for a global
interpreter to be set for all tasks.